requirements.txt
arrow
attrs
certifi
charset-normalizer
cmake
Cython
distro
docker
docstring-parser
filelock
fsspec
hypothesis
idna
iopath
Jinja2
MarkupSafe
mypy-extensions
ninja
numpy
packaging
pandas
portalocker
pyarrow
pyDeprecate
pyparsing
pyre-extensions
python-dateutil
pytz
PyYAML
requests
scikit-build
six
sortedcontainers
tabulate
torchmetrics
torchx-nightly
tqdm
typing
typing-inspect
typing_extensions
urllib3
websocket-client
rich
